apply formatting

+++ key.util/src/test/java/org/key_project/util/testcase/java/IOUtilTest.java
@@ -156,9 +156,12 @@ public class IOUtilTest {
     /**
      * Performs a test step of {@link #testUnifyLineBreaks()}.
      *
-     * @param toTest The {@link String} to test.
-     * @param expected The expected result.
-     * @throws IOException Occurred Exception.
+     * @param toTest
+     *        The {@link String} to test.
+     * @param expected
+     *        The expected result.
+     * @throws IOException
+     *         Occurred Exception.
      */
     protected void doTestUnifyLineBreaks(String toTest, String expected) throws IOException {
         ByteArrayInputStream in =
@@ -182,7 +185,8 @@ public class IOUtilTest {
         /**
          * Constructor.
          *
-         * @param text The fixed text.
+         * @param text
+         *        The fixed text.
          */
         public TextInputStream(String text) {
             super(text.getBytes(StandardCharsets.UTF_8));
@@ -430,10 +434,14 @@ public class IOUtilTest {
     /**
      * Executes a test for {@link #testLineInformationNormalizeColumn()}.
      *
-     * @param text The text to use.
-     * @param tabWidth The tab width to use.
-     * @param expectedIndices The expected normalized indices.
-     * @throws IOException Occurred Exception.
+     * @param text
+     *        The text to use.
+     * @param tabWidth
+     *        The tab width to use.
+     * @param expectedIndices
+     *        The expected normalized indices.
+     * @throws IOException
+     *         Occurred Exception.
      */
     protected void doTestLineInformationNormalizeColumn(String text, int tabWidth,
             int[] expectedIndices) throws IOException {
@@ -459,7 +467,7 @@ public class IOUtilTest {
     public void testComputeLineInformation_File() throws IOException {
         // Get test file
         File textFile = new File(HelperClassForUtilityTests.RESOURCE_DIRECTORY + File.separator
-            + "lineIndicesTest" + File.separator + "Text.txt");
+                + "lineIndicesTest" + File.separator + "Text.txt");
         assertTrue(textFile.isFile(), "File '" + textFile + "' does not exist.");
         // Test null
         assertLineInformation((File) null);
@@ -484,11 +492,15 @@ public class IOUtilTest {
      * possible to commit/checkout the test data files directly.
      * </p>
      *
-     * @param source The {@link File} with the source text.
-     * @param newFileName The new file name.
-     * @param lineBreak The line break to use.
+     * @param source
+     *        The {@link File} with the source text.
+     * @param newFileName
+     *        The new file name.
+     * @param lineBreak
+     *        The line break to use.
      * @return The created {@link File} with the same text but with new line breaks.
-     * @throws IOException Occurred Exception
+     * @throws IOException
+     *         Occurred Exception
      */
     protected File convertTextFile(File source, String newFileName, String lineBreak)
             throws IOException {
@@ -501,10 +513,7 @@ public class IOUtilTest {
                     new BufferedReader(new InputStreamReader(new FileInputStream(source),
                         StandardCharsets.UTF_8))) {
             String line;
-            while ((line = reader.readLine()) != null) {
-                writer.write(line);
-                writer.write(lineBreak);
-            }
+            while ((line = reader.readLine()) != null) { writer.write(line); writer.write(lineBreak); }
             String newText = writer.toString();
             // Create new file
             File target = new File(source.getParentFile(), newFileName);
@@ -518,9 +527,12 @@ public class IOUtilTest {
     /**
      * Makes sure that for the given text the correct line start indices are computed.
      *
-     * @param file The text to test.
-     * @param expectedIndices The expected line indices.
-     * @throws IOException Occurred Exception.
+     * @param file
+     *        The text to test.
+     * @param expectedIndices
+     *        The expected line indices.
+     * @throws IOException
+     *         Occurred Exception.
      */
     protected void assertLineInformation(File file, int... expectedIndices) throws IOException {
         LineInformation[] result = IOUtil.computeLineInformation(file);
@@ -546,8 +558,10 @@ public class IOUtilTest {
      * Executes the tests for {@link #testComputeLineInformation_InputStream()} with the given line
      * break sign.
      *
-     * @param newLine The line break sign to use.
-     * @throws IOException Occurred Exception.
+     * @param newLine
+     *        The line break sign to use.
+     * @throws IOException
+     *         Occurred Exception.
      */
     protected void doTestComputeLineInformation_InputStream(String newLine) throws IOException {
         // Test null
@@ -586,9 +600,12 @@ public class IOUtilTest {
     /**
      * Constructs a text for the given lines and tests the computed start line indices.
      *
-     * @param newLine The new line sign to use.
-     * @param textLines The lines of text.
-     * @throws IOException Occurred Exception.
+     * @param newLine
+     *        The new line sign to use.
+     * @param textLines
+     *        The lines of text.
+     * @throws IOException
+     *         Occurred Exception.
      */
     protected void assertLineInformation(String newLine, String... textLines) throws IOException {
         if (textLines != null) {
@@ -599,11 +616,7 @@ public class IOUtilTest {
                 // Compute tabs
                 List<Integer> tabIndices = new LinkedList<>();
                 char[] lineChars = textLines[i].toCharArray();
-                for (int j = 0; j < lineChars.length; j++) {
-                    if ('\t' == lineChars[j]) {
-                        tabIndices.add(j);
-                    }
-                }
+                for (int j = 0; j < lineChars.length; j++) { if ('\t' == lineChars[j]) { tabIndices.add(j); } }
                 // Compute line
                 expectedInfos[i] = new LineInformation(lastIndex, tabIndices);
                 sb.append(textLines[i]);
@@ -622,9 +635,12 @@ public class IOUtilTest {
     /**
      * Makes sure that for the given text the correct line start indices are computed.
      *
-     * @param text The text to test.
-     * @param expectedInfos The expected line informations.
-     * @throws IOException Occurred Exception.
+     * @param text
+     *        The text to test.
+     * @param expectedInfos
+     *        The expected line informations.
+     * @throws IOException
+     *         Occurred Exception.
      */
     protected void assertLineInformation(String text, LineInformation... expectedInfos)
             throws IOException {
@@ -668,9 +684,7 @@ public class IOUtilTest {
             IOUtil.writeTo(new FileOutputStream(tempFile), content);
             assertEquals(content, IOUtil.readFrom(new FileInputStream(tempFile)));
         } finally {
-            if (tempFile != null) {
-                tempFile.delete();
-            }
+            if (tempFile != null) { tempFile.delete(); }
         }
     }
 
@@ -689,10 +703,13 @@ public class IOUtilTest {
     /**
      * Performs test steps of {@link #testWriteTo_Charstet()}.
      *
-     * @param text The text to write.
-     * @param encoding The encoding to use.
+     * @param text
+     *        The text to write.
+     * @param encoding
+     *        The encoding to use.
      * @return The written bytes.
-     * @throws Exception Occurred Exception.
+     * @throws Exception
+     *         Occurred Exception.
      */
     protected byte[] doWriteCharsetAsXmlTest(String text, Charset encoding) throws Exception {
         // Create XML
@@ -860,8 +877,10 @@ public class IOUtilTest {
     /**
      * Executes the assertions for {@link #testCopy()}.
      *
-     * @param text The text to check.
-     * @throws IOException Occurred Exception.
+     * @param text
+     *        The text to check.
+     * @throws IOException
+     *         Occurred Exception.
      */
     protected void doTestCopy(String text) throws IOException {
         if (text != null) {
@@ -871,9 +890,7 @@ public class IOUtilTest {
             assertTrue(IOUtil.copy(in, out));
             byte[] outBytes = out.toByteArray();
             assertEquals(inBytes.length, outBytes.length);
-            for (int i = 0; i < inBytes.length; i++) {
-                assertEquals(inBytes[i], outBytes[i]);
-            }
+            for (int i = 0; i < inBytes.length; i++) { assertEquals(inBytes[i], outBytes[i]); }
         } else {
             assertFalse(IOUtil.copy(null, new ByteArrayOutputStream()));
         }
@@ -900,7 +917,8 @@ public class IOUtilTest {
     /**
      * Tests {@link IOUtil#toURI(java.net.URL)}
      *
-     * @throws MalformedURLException Occurred Exception
+     * @throws MalformedURLException
+     *         Occurred Exception
      * @see IOUtil#toURI(java.net.URL)
      */
     @Test
@@ -931,7 +949,8 @@ public class IOUtilTest {
     /**
      * Tests {@link IOUtil#toFile(URL)}
      *
-     * @throws MalformedURLException Occurred Exception
+     * @throws MalformedURLException
+     *         Occurred Exception
      * @see IOUtil#toFile(URL)
      */
     @Test
@@ -949,7 +968,8 @@ public class IOUtilTest {
     /**
      * Tests {@link IOUtil#toFileString(URL)}
      *
-     * @throws MalformedURLException Occurred Exception
+     * @throws MalformedURLException
+     *         Occurred Exception
      * @see IOUtil#toFileString(URL)
      */
     @Test()
